Trust QR advertises agent DID and FQN entitlements
The --trust QR now carries the agent's own identity, not the device's:
start_agent_server loads/creates a persisted agent keypair (Task 1's
get_agent_keypair/store_agent_keypair slot) alongside the device
keypair, and the AgentDescriptor/QR is built from that agent public
key. Entitlements move from the old agent.capability.* strings to
attribute FQNs (https://arkavo.ai/attr/...), the vocabulary authnz-rs
delegates from the human's stored entitlements, sourced from a new
AgentConfig.entitlements field (parsed from AGENTS.md front-matter,
defaulting to DEFAULT_AGENT_ENTITLEMENTS). The agent's DID is printed
under the QR so the human can cross-check it against the phone app.

AgentConfig/McpServerConfig/parse_agents_config and their parsing
helpers move to a new commands/agent_config.rs (agent.rs was already
over 400 non-test lines), re-exported via pub(crate) use.

Also: arkavo-registration's with_entitlements doc comment and URL
test updated to the FQN vocabulary; a DEVICE-007 test added asserting
the agent DID differs from the device DID (Task 1 only checked raw
key bytes); qr-registration.spec.yaml's QREG-003 updated to describe
FQN entitlements; index.yaml's stale total_scenarios counter bumped
to 752.
