+++ crates/arkavo-cli/src/commands/mod.rs
@@ -1,4 +1,5 @@
 pub mod agent;
+pub mod agent_config;
 pub mod chat;
 pub mod dataflow;
 pub mod mesh;
