+++ crates/arkavo-device-identity/Cargo.toml
@@ -28,6 +28,7 @@ dirs = "5.0"
 [dev-dependencies]
 serde_json.workspace = true
 arkavo-test-macros = { path = "../arkavo-test-macros" }
+arkavo-crypto = { path = "../arkavo-crypto" }
 
 [features]
 default = []
